[refactor] moving classes to core package
