+++ src/main/java/fr/inria/coming/core/filter/commitmessage/KeyWordsTitleFilter.java
@@ -1,5 +1,6 @@
-package fr.inria.coming.core.filter;
+package fr.inria.coming.core.filter.commitmessage;
 
+import fr.inria.coming.core.filter.AbstractChainedFilter;
 import fr.inria.coming.core.interfaces.Commit;
 import fr.inria.coming.core.interfaces.IFilter;
 
@@ -9,7 +10,7 @@ import fr.inria.coming.core.interfaces.IFilter;
  * A filter to search keywords ONLY in the first line of a commit message
  *
  */
-public class KeyWordsTitleFilter extends AbstractFilter {
+public class KeyWordsTitleFilter extends AbstractChainedFilter {
 	
 	
 	protected String [] predicates;
