+++ src/test/java/fr/inria/coming/spoon/patterns/ScenariosTest.java
@@ -13,11 +13,11 @@ import org.junit.Test;
 
 import com.github.gumtreediff.matchers.Matcher;
 
-import fr.inria.coming.changeminer.analyzer.Parameters;
 import fr.inria.coming.changeminer.entity.GranuralityType;
 import fr.inria.coming.changeminer.util.ConsoleOutput;
 import fr.inria.coming.changeminer.util.Scenarios;
 import fr.inria.coming.changeminer.util.XMLOutput;
+import fr.inria.coming.core.Parameters;
 import fr.inria.coming.core.interfaces.Commit;
 import gumtree.spoon.diff.operations.Operation;
